update admin manage user

+++ src/main/java/ku/cs/palmoilmnger/service/UserService.java
@@ -4,6 +4,7 @@ import ku.cs.palmoilmnger.entity.User;
 import ku.cs.palmoilmnger.exception.UserException;
 import ku.cs.palmoilmnger.model.ChangePassword;
 import ku.cs.palmoilmnger.repository.UserRepository;
+import ku.cs.palmoilmnger.repository.WorkRoundRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -14,6 +15,8 @@ import java.util.List;
 public class UserService {
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private WorkRoundRepository workRoundRepository;
 
     @Autowired
     private PasswordEncoder passwordEncoder;
@@ -21,17 +24,17 @@ public class UserService {
     @Autowired
     private WorkRoundService workRoundService;
 
-    private boolean isUsernameNotAvailable(String username){
+    private boolean isUsernameNotAvailable(String username) {
         return userRepository.findByUsername(username) == null;
     }
 
-    private boolean isNameNotAvailable(String name){
+    private boolean isNameNotAvailable(String name) {
         return userRepository.findByName(name) == null;
     }
 
     public void createManager(User user) throws UserException {
-        if(!isNameNotAvailable(user.getName())) throw new UserException("ชื่อมีในระบบแล้ว");
-        if(!isUsernameNotAvailable(user.getUsername())) throw new UserException("ชื่อผู้ใช้มีในระบบแล้ว");
+        if (!isNameNotAvailable(user.getName())) throw new UserException("ชื่อมีในระบบแล้ว");
+        if (!isUsernameNotAvailable(user.getUsername())) throw new UserException("ชื่อผู้ใช้มีในระบบแล้ว");
 
         User record = new User();
         record.setName(user.getName());
@@ -44,24 +47,25 @@ public class UserService {
         userRepository.save(record);
     }
 
-    public User getManager(String username){
+    public User getManager(String username) {
         return userRepository.findByUsername(username);
     }
 
-    public List<User> getAllUsersRole(String role){
+    public List<User> getAllUsersRole(String role) {
         return userRepository.findByRole(role);
     }
 
-    public List<User> getAllUsers(){
+    public List<User> getAllUsers() {
         return userRepository.findAll();
     }
 
     // changePassword method
     public void changePassword(ChangePassword changePassword) throws UserException {
-        if(!changePassword.getPassword().equals(changePassword.getConfirmPassword())) throw new UserException("รหัสผ่านไม่ตรงกัน");
+        if (!changePassword.getPassword().equals(changePassword.getConfirmPassword()))
+            throw new UserException("รหัสผ่านไม่ตรงกัน");
 
         User record = userRepository.findByUsername(changePassword.getUsername());
-        if(record == null) throw new UserException("ไม่พบผู้ใช้ในระบบ");
+        if (record == null) throw new UserException("ไม่พบผู้ใช้ในระบบ");
 
         String hashed = passwordEncoder.encode(changePassword.getPassword());
         record.setPassword(hashed);
@@ -69,9 +73,11 @@ public class UserService {
     }
 
     // Delete User method
-    public void deleteUser(String username){
+    public void deleteUser(String username) throws UserException {
         User user = this.getManager(username);
-
+        if (!workRoundRepository.findByUser(user).isEmpty()) {
+            throw new UserException("มีรอบที่ผู้ใช้ทำงานอยู่");
+        }
         userRepository.delete(user);
     }
 }
